requirements.txt
azure-identity
azure-keyvault-secrets
azure-mgmt-resource
azure-mgmt-subscription
msgraph-core
requests